Initial stab at integration into ItemUtil
This is not complete and I am not happy with how I force-fit the type
specific behavior. Ideally we derive a specific converter based on the
generic type. This should be pretty achievable by having a dictionary
with mapping from type -> conveter or cool logic for extrapolating the
type from the base DredgeTypeConverter which is based on the same one.